backend/requirements.txt
Django==2.1.2
django-crispy-forms==1.7.2
pytz==2018.7
rest_framework_pagination
drf-yasg
django-extensions
django-autoslug
Pillow
djangorestframework-simplejwt
django-cors-headers
PyJWT==1.7.1